+++ app/src/test/java/org/simple/clinic/home/overdue/search/OverdueSearchEffectHandlerTest.kt
@@ -205,4 +205,15 @@ class OverdueSearchEffectHandlerTest {
     verify(uiActions).openSelectDownloadFormatDialog(emptySet())
     verifyNoMoreInteractions(uiActions)
   }
+
+  @Test
+  fun `when show no internet connection dialog effect is received, then show no internet connection dialog`() {
+    // when
+    effectHandlerTestCase.dispatch(ShowNoActiveNetworkConnectionDialog)
+
+    // then
+    verify(uiActions).showNoActiveNetworkConnectionDialog()
+    effectHandlerTestCase.assertNoOutgoingEvents()
+    verifyNoMoreInteractions(uiActions)
+  }
 }