+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Remove extra margin start and end for overdue search button
 - Download overdue search results in CSV format when PDF cannot be generated
 - Open select overdue download format when PDF can be generated in `OverdueSearchScreen`
+- Show no active connection dialog when download button is clicked and network is not connected in `OverdueSearchScreen`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-11-8325